pyproject.toml
[tool.poetry]
name = "beginner.py"
version = "0.1.1-alpha.0"
description = ""
authors = ["Zech Zimmerman <hi@zech.codes>"]

[tool.poetry.dependencies]
python = "^3.12"
PyYAML = "^5.4.1"
SQLAlchemy = "^1.4.1"
psycopg2-binary = "^2.8.6"
arrow = "^1.1.0"
nextcord = "^2.6.0"
"dippy.bot" = "^0.1.1-alpha.28"

[[tool.poetry.source]]
name = "pypi"
url = "https://pypi.org/simple"
priority = "primary"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
